% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{getxl}
\alias{getxl}
\title{getxl}
\usage{
getxl(input, output, session)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}
}
\description{
Convert excel sheets to dataframes, display in DT, and return the dataframes
}
\details{
This is a shiny module that presents a file picker UI to get an Excel file name, and reads the Excel sheets using readxl package and returns the resulting sheet(s) as a vector and data in dataframe(s).
}
\seealso{
See xlex for examples
}
