% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{xldisp}
\alias{xldisp}
\title{xldisp}
\usage{
xldisp(input, output, session, xlfile, skiprows)
}
\arguments{
\item{input}{is shiny input variable}

\item{output}{is shiny output variable}

\item{session}{is shiny session variable}

\item{xlfile}{is the name of the xl file being got}

\item{skiprows}{is number of rows to skip in the excel file}
}
\description{
Server function to display excel data as DT
}
