% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.R
\name{xldispUI}
\alias{xldispUI}
\title{xldispUI}
\usage{
xldispUI(id, xlfile)
}
\arguments{
\item{id}{is caller's id}

\item{xlfile}{is the name of the xl file being got}
}
\description{
UI to display excel sheets read as individual data tables
}
